+++ src/file_io.cpp
@@ -38,14 +38,14 @@ unsigned char *read_all(const char *file_name, bool text_mode, unsigned int mini
     return file_data;
 }
 
-patchfile write_all(unsigned char* data, std::string_view file_name, unsigned int size) {
+patchfile write_all(const unsigned char* data, std::string_view file_name, unsigned int size) {
     patchfile file{std::string{file_name}, static_cast<patchfile::openflags>(std::ios::out | std::ios::binary)};
     file.fwrite(data, size);
     file.close();
     return file;
 }
 
-patchfile write_all(unsigned char* data, std::string_view dir, std::string_view file_name, unsigned int size) {
+patchfile write_all(const unsigned char* data, std::string_view dir, std::string_view file_name, unsigned int size) {
     std::string fullpath{dir};
 	fullpath += file_name;
     return write_all(data, fullpath, size);